@@ -10,7 +10,7 @@
 
 	<groupId>com.thekayani</groupId>
 	<artifactId>hklib</artifactId>
-	<version>1.2.2</version>
+	<version>1.3.0</version>
 	<name>${project.groupId}:${project.artifactId}</name>
 	<description>
 		This library contains various helper methods, classes, and features to enable rapid application development.
@@ -24,8 +24,8 @@
 	</licenses>
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<maven.compiler.source>1.7</maven.compiler.source>
-		<maven.compiler.target>1.7</maven.compiler.target>
+		<maven.compiler.source>1.8</maven.compiler.source>
+		<maven.compiler.target>1.8</maven.compiler.target>
 	</properties>
 	<developers>
 		<developer>
@@ -52,7 +52,7 @@
 	<dependencies>
 		<dependency>
 			<groupId>org.jetbrains</groupId>
-			<artifactId>annotations-java5</artifactId>
+			<artifactId>annotations</artifactId>
 			<version>23.0.0</version>
 			<scope>provided</scope>
 		</dependency>
@@ -70,8 +70,8 @@
 				<artifactId>maven-compiler-plugin</artifactId>
 				<version>3.8.1</version>
 				<configuration>
-					<source>1.7</source>
-					<target>1.7</target>
+					<source>1.8</source>
+					<target>1.8</target>
 					<excludes>
 						<exclude>com/hk/Main.java</exclude>
 					</excludes>
