@@ -3,6 +3,7 @@ package com.hk.io;
 import java.io.Writer;
 
 import com.hk.util.Requirements;
+import org.jetbrains.annotations.NotNull;
 
 /**
  * <p>StringBuilderWriter class.</p>
@@ -36,14 +37,14 @@ public class StringBuilderWriter extends Writer
 	 *
 	 * @param sb a {@link java.lang.StringBuilder} object
 	 */
-	public StringBuilderWriter(StringBuilder sb)
+	public StringBuilderWriter(@NotNull StringBuilder sb)
 	{
 		this.sb = Requirements.requireNotNull(sb);
 	}
 
 	/** {@inheritDoc} */
 	@Override
-	public void write(char[] cbuf, int off, int len)
+	public void write(char @NotNull [] cbuf, int off, int len)
 	{
 		sb.append(cbuf, off, len);
 	}
@@ -67,6 +68,7 @@ public class StringBuilderWriter extends Writer
 	 *
 	 * @return a {@link java.lang.StringBuilder} object
 	 */
+	@NotNull
 	public StringBuilder getBuilder()
 	{
 		return sb;
@@ -77,6 +79,7 @@ public class StringBuilderWriter extends Writer
 	 *
 	 * @return a {@link java.lang.String} object
 	 */
+	@NotNull
 	public String toString()
 	{
 		return sb.toString();
