@@ -1,5 +1,7 @@
 package com.hk.json;
 
+import org.jetbrains.annotations.NotNull;
+
 /**
  * <p>Abstract JsonValue class.</p>
  *
@@ -25,6 +27,7 @@ public abstract class JsonValue
 	 *
 	 * @return a {@link com.hk.json.JsonType} object
 	 */
+	@NotNull
 	public abstract JsonType getType();
 
 	/**
@@ -32,6 +35,7 @@ public abstract class JsonValue
 	 *
 	 * @return a {@link com.hk.json.JsonObject} object
 	 */
+	@NotNull
 	public JsonObject getObject()
 	{
 		throw new IllegalStateException("not an object");
@@ -52,6 +56,7 @@ public abstract class JsonValue
 	 *
 	 * @return a {@link com.hk.json.JsonArray} object
 	 */
+	@NotNull
 	public JsonArray getArray()
 	{
 		throw new IllegalStateException("not an array");
@@ -72,6 +77,7 @@ public abstract class JsonValue
 	 *
 	 * @return a {@link java.lang.String} object
 	 */
+	@NotNull
 	public String getString()
 	{
 		throw new IllegalStateException("not a string");
@@ -92,6 +98,7 @@ public abstract class JsonValue
 	 *
 	 * @return a {@link java.lang.Number} object
 	 */
+	@NotNull
 	public Number getNumber()
 	{
 		throw new IllegalStateException("not a number");
@@ -125,7 +132,7 @@ public abstract class JsonValue
 	 * @param <T> a T class
 	 * @return a boolean
 	 */
-	public <T> boolean is(Class<T> cls, JsonAdapter<?> adapter)
+	public <T> boolean is(@NotNull Class<T> cls, @NotNull JsonAdapter<?> adapter)
 	{
 		return adapter.getObjClass().isAssignableFrom(cls);
 	}
@@ -137,7 +144,7 @@ public abstract class JsonValue
 	 * @param <T> a T class
 	 * @return a boolean
 	 */
-	public <T> boolean is(Class<T> cls)
+	public <T> boolean is(@NotNull Class<T> cls)
 	{
 		for(JsonAdapter<?> adapter : Json.globalAdapters)
 		{
@@ -155,7 +162,7 @@ public abstract class JsonValue
 	 * @return a T object
 	 * @throws com.hk.json.JsonAdaptationException if any.
 	 */
-	public <T> T get(JsonAdapter<T> adapter) throws JsonAdaptationException
+	public <T> T get(@NotNull JsonAdapter<T> adapter) throws JsonAdaptationException
 	{
 		return adapter.fromJson(this);
 	}
@@ -169,7 +176,7 @@ public abstract class JsonValue
 	 * @throws com.hk.json.JsonAdaptationException if any.
 	 */
 	@SuppressWarnings("unchecked")
-	public <T> T get(Class<T> cls) throws JsonAdaptationException
+	public <T> T get(@NotNull Class<T> cls) throws JsonAdaptationException
 	{
 		for(JsonAdapter<?> adapter : Json.globalAdapters)
 		{
@@ -194,6 +201,7 @@ public abstract class JsonValue
 	 *
 	 * @return a {@link java.lang.String} object
 	 */
+	@NotNull
 	public String toString()
 	{
 		return Json.writePretty(this);
