@@ -1,7 +1,8 @@
 package com.hk.lua;
 
-import com.hk.func.BiConsumer;
 import com.hk.lua.Lua.LuaMethod;
+
+import java.util.function.BiConsumer;
 
 /**
  * <p>LuaLibraryTable class.</p>
@@ -162,6 +163,6 @@ public enum LuaLibraryTable implements BiConsumer<Environment, LuaObject>, LuaMe
 	{
 		String name = toString();
 		if(name != null && !name.trim().isEmpty())
-			table.rawSet(new LuaString(name), Lua.newFunc(this));
+			table.rawSet(new LuaString(name), Lua.newMethod(this));
 	}
 }
